# soroban-rpc: further DB querying performance tweaks